Add support for `uint` in PersistenceManager and use `uint` for playlist settings as these should never be negtive

+++ src/Modules/Playlist/Models/PlaylistSettings.cs
@@ -6,11 +6,11 @@ namespace Whitestone.SegnoSharp.Modules.Playlist.Models
     {
         [Persist]
         [DefaultValue(3)]
-        public int MinimumNumberOfSongs { get; set; }
+        public ushort MinimumNumberOfSongs { get; set; }
 
         [Persist]
         [DefaultValue(15)]
-        public int MinimumTotalDuration { get; set; }
+        public ushort MinimumTotalDuration { get; set; }
 
     }
 }
